# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "My API",
      "description": "This is a sample API documentation",
      "version": "1.0.0",
      "contact": {
        "name": "SilkyWay API Support",
        "url": "http://www.silkyway.vercel.app",
        "email": "silkyway.enterprise@gmail.com"
      }
    },
    "servers": [
      {
        "url": "http://localhost:8080",
        "description": "Local server"
      },
      {
        "url": "https://silkyway.vercel.app",
        "description": "Production server"
      }
    ],
...
    "paths": {
      "/api/products": {
        "get": {
          "tags": ["Products"],
          "summary": "Get list of products",
          "responses": {
            "200": {
              "description": "A list of products",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad request"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        },
        "post": {
          "tags": ["Products"],
          "summary": "Create a new product",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "Name of the product",
                      "example": "Sample Product"
                    },
                    "description": {
                      "type": "string",
                      "description": "Description of the product",
                      "example": "This is a sample product."
                    },
                    "price": {
                      "type": "number",
                      "description": "Price of the product",
                      "example": 19.99
                    },
                    "stock": {
                      "type": "number",
                      "description": "Stock of the product",
                      "example": 100
                    },
                    "images": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "url": {
                          "type": "string",
                          "description": "URLs of product images",
                          "example": "http://example.com/image1.jpg"
                        },
                        "alt": {
                          "type": "string",
                          "description": "Image description",
                          "example": "Image description"
                        }
                      }
                    },
                    "description": "Images of the product"
                  },
                  "categoryID": {
                    "type": "string",
                    "description": "ID of the product category",
                    "example": "60d6e88f53112345678d960c"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "description": "Tags associated with the product",
                      "example": "new"
                    },
                    "description": "Tags of the product"
                  }
                },
                "required": ["name", "price", "images"]
              }
            }
          },
          "required": true
        },
          "responses": {
            "201": {
              "description": "Product created",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            },
            "400": {
              "description": "Bad request"
            },
            "409": {
              "description": "Conflict - Product with the same name already exists"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
...
      "api/auth/login": {
        "post": {
          "tags": ["Authentication"],
          "summary": "Login an existing user",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string",
                      "format": "email",
                      "example": "user@example.com"
                    },
                    "password": {
                      "type": "string",
                      "example": "password123"
                    }
                  },
                  "required": ["email", "password"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "User successfully logged in",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "user": {
                        "$ref": "#/components/schemas/User"
                      },
                      "token": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad request - Invalid email/password"
            },
            "401": {
              "description": "Unauthorized - Incorrect email/password"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
      "api/auth/register": {
        "post": {
          "tags": ["Users"],
          "summary": "Register a new user",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRegister"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "User successfully registered",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/UserRegister"
                  }
                }
              }
            },
            "400": {
              "description": "Bad request - Email address already in use or invalid data format"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
        "Product": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string",
              "required": true,
              "trim": true,
              "example": "Sample Product"
            },
            "description": {
              "type": "string",
              "required": true,
              "example": "This is a sample product description"
            },
            "price": {
              "type": "number",
              "example": 19.99
            },
            "categoryID": {
              "type": "number",
              "example": 1
            },
            "tags": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "stock": {
              "type": "number",
              "example": 100
            },
            "images": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "url": {
                    "type": "string",
                    "example": "http://example.com/image1.jpg"
                  },
                  "alt": {
                    "type": "string",
                    "example": "Image description"
                  }
                }
              }
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "example": "2023-06-01T00:00:00Z"
            },
            "updated_at": {
              "type": "string",
              "format": "date-time",
              "example": "2023-06-01T00:00:00Z"
            }
          }
        },
...
        "User": {
          "type": "object",
          "properties": {
            "email": {
              "type": "string",
              "required": true,
              "unique": true,
              "example": "user@example.com"
            },
            "name": {
              "type": "string",
              "required": true,
              "trim": true,
              "example": "John Doe"
            },
            "password": {
              "type": "string",
              "required": true,
              "minlength": 6,
              "example": "password123"
            }
          }
        },
        "UserRegister": {
          "type": "object",
          "properties": {
            "email": {
              "type": "string",
              "required": true,
              "unique": true,
              "example": "user@example.com"
            },
            "password": {
              "type": "string",
              "required": true,
              "minlength": 6,
              "example": "password123"
            },
            "firstname": {
              "type": "string",
              "required": true,
              "trim": true,
              "example": "John"
            },
            "lastname": {
              "type": "string",
              "required": true,
              "trim": true,
              "example": "Doe"
            },
            "address": {
              "type": "string",
              "required": true,
              "trim": true,
              "example": "example@hotmail.com"
            }
          }
        }
      }
    }
  }